Moved parser/generator to the detail namespace.
